legacy.backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dsa110-contimg"
version = "0.1.0"
description = "DSA-110 Continuum Imaging Pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "DSA-110 Team"}
]
keywords = ["radio astronomy", "imaging", "calibration", "CASA", "DSA-110"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

# Core runtime dependencies
# Note: CASA packages (casatools, casatasks, casacore) are typically installed
# via conda, not pip, so they're listed as optional dependencies
dependencies = [
    "astropy>=5.0.0",
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "pandas>=1.5.0",
    "h5py>=3.7.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "psutil>=5.9.0",
    "watchdog>=3.0.0",
    "typing-extensions>=4.0.0",
    "pyyaml>=6.0",
    "uncertainties>=3.2.0",
    "healpy>=1.16.0",  # For HEALPix sky maps (Mollweide projection)
    "pygdsm>=1.5.0",   # For Global Sky Model generation
]

# Optional dependencies for CASA (typically installed via conda)
[project.optional-dependencies]
casa = [
    # These are typically installed via conda, but listed here for reference
    # "casatools",
    # "casatasks",
    # "casaconfig",
    # "python-casacore",
]
qa = [
    # Optional QA tools for visualization and inspection
    "ragavi>=0.5.2",  # Radio Astronomy Gain and Visibility Inspector (requires codex-africanus>=0.4.3)
    # Note: codex-africanus is a transitive dependency of ragavi
    # Version 0.4.3+ supports numba>=0.60 (compatible with numba 0.62+)
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "pytest-timeout>=2.1.0",
    "pytest-mock>=3.11.0",
    "coverage>=7.2.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pylint>=3.0.0",
    "pyflakes>=3.0.0",
    "mypy>=1.4.0",
    "pytest-benchmark>=4.0.0",
    "memory-profiler>=0.61.0",
    "httpx>=0.25.0",
]
all = [
    "dsa110-contimg[dev]",
]

[project.scripts]
# Main CLI entry points
dsa110-convert = "dsa110_contimg.conversion.cli:main"
dsa110-calibrate = "dsa110_contimg.calibration.cli:main"
dsa110-image = "dsa110_contimg.imaging.cli:main"
dsa110-mosaic = "dsa110_contimg.mosaic.cli:main"
dsa110-photometry = "dsa110_contimg.photometry.cli:main"
dsa110-registry = "dsa110_contimg.database.registry_cli:main"
dsa110-streaming = "dsa110_contimg.conversion.streaming.streaming_converter:main"
# Performance benchmarking
dsa110-benchmark = "dsa110_contimg.benchmark.cli:main"
# DOCUMENTATION FIX: Add CLI entry points referenced in documentation
dsa110-contimg-convert-stream = "dsa110_contimg.conversion.streaming.streaming_converter:main"
dsa110-contimg-streaming-mosaic = "dsa110_contimg.mosaic.streaming_mosaic:main"

[tool.setuptools]
# Package discovery - find all packages in src/
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["dsa110_contimg*"]
exclude = ["tests*", "*.tests*", "*.tests.*", "tests.*"]

# Black code formatter configuration
[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | archive
)/
'''

# isort import sorting
[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "scripts", "tests"]
skip_glob = ["archive/*", "*.pyc"]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (require external resources)",
    "slow: Slow tests (may take minutes)",
    "casa: Tests requiring CASA environment",
    "synthetic: Tests using synthetic data",
    "e2e: End-to-end tests",
    "smoke: Smoke tests (quick sanity checks)",
    "science: Science validation tests",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/archive/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
# Coverage thresholds: 80% for critical modules, 60% overall
fail_under = 60
precision = 2
show_missing = true
skip_covered = false

# MyPy type checking
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

# Pylint configuration
[tool.pylint.main]
# CRITICAL: Use Python 3.11 (casa6) - NOT Python 3.6
# This ensures Pylint analyzes code with the correct Python version
py-version=3.11
# Suppress false positives for astropy.units dynamic attributes
# Note: astropy.units dynamically creates unit attributes (deg, arcsec, etc.)
# at import time, which pylint's static analysis cannot detect
extension-pkg-allow-list=["astropy"]

# Generated members for dynamic attributes
# Astropy units creates attributes like deg, arcsec, hourangle, m, arcmin dynamically
# CASA tools create methods dynamically on coordinatesystem objects
# FITS HDUList behaves like a list but pylint doesn't recognize indexing
generated-members = [
    "astropy.units.*",
    "deg", "arcsec", "hourangle", "arcmin", "m", "cm", "dex",
    "u.deg", "u.arcsec", "u.hourangle", "u.m", "u.arcmin", "u.cm",
]

[tool.pylint.format]
max-line-length = 120

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]

[tool.pylint."messages control"]
disable = [
    "no-member",  # Astropy, CASA, FITS have dynamic attributes
    "import-error",  # Optional modules like pipeline_hooks may not exist
    "no-name-in-module",  # CASA modules have dynamic exports
    "unexpected-keyword-arg",  # CASA APIs may have version-specific parameters
]
# Note: We cannot selectively disable no-member for just astropy.units
# The astropy.units errors are false positives - units exist at runtime
# but are created dynamically, so pylint's static analysis can't see them.
# Consider adding # pylint: disable=no-member to specific lines using astropy.units
# or use a more targeted approach with inline suppressions
# CASA modules are not typed
[[tool.mypy.overrides]]
module = [
    "casatools.*",
    "casatasks.*",
    "casacore.*",
    "pyuvdata.*",
]
ignore_missing_imports = true

